backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
eth-account==0.9.0
web3==6.11.0
python-dotenv==1.0.0
setuptools>=65.0.0
